src/KhashChat/handlers/advice.py
from telegram.ext import CommandHandler

def advice(update, context):
    # Implement your financial advice logic here
    advice_text = "Here's some financial advice for you..."
    context.bot.send_message(chat_id=update.effective_chat.id, text=advice_text)

def advice_handler():
    return CommandHandler('advice', advice)
